pyproject.toml
[project]
name = "dreamwright-workspace"
version = "0.1.0"
description = "AI-powered webtoon and short-form drama production - Monorepo"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dreamwright-core-schemas",
    "dreamwright-services",
    "dreamwright-generators",
    "dreamwright-storage",
    "dreamwright-gemini-client",
    "dreamwright-cli",
    "dreamwright-api",
    "dreamwright-viewer",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]
docs = [
    "pyyaml>=6.0.0",
]

# No build-system for workspace root - it's just a coordinator

[tool.uv.workspace]
members = [
    "apps/*",
    "packages/*",
]

[tool.uv.sources]
dreamwright-core-schemas = { workspace = true }
dreamwright-services = { workspace = true }
dreamwright-generators = { workspace = true }
dreamwright-storage = { workspace = true }
dreamwright-gemini-client = { workspace = true }
dreamwright-cli = { workspace = true }
dreamwright-api = { workspace = true }
dreamwright-viewer = { workspace = true }

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
